refactor compile() to abstract parent class
